chore: Move NavigationData to home provider
